+++ tests/unit/data/utils/test_utils.py
@@ -5,12 +5,13 @@
 
 from __future__ import annotations
 
+from collections import defaultdict
 from unittest.mock import MagicMock
 
 import cv2
 import numpy as np
 import pytest
-from datumaro.components.annotation import Bbox
+from datumaro.components.annotation import Bbox, Label
 from datumaro.components.dataset import Dataset as DmDataset
 from datumaro.components.dataset_base import DatasetItem
 from datumaro.components.media import Image
@@ -21,6 +22,9 @@ from otx.data.utils.utils import (
     compute_robust_dataset_statistics,
     compute_robust_scale_statistics,
     compute_robust_statistics,
+    get_adaptive_num_workers,
+    get_idx_list_per_classes,
+    import_object_from_module,
 )
 
 
@@ -185,3 +189,79 @@ def test_adapt_input_size_to_dataset(mocker):
     }
     input_size = adapt_input_size_to_dataset(dataset=MagicMock())
     assert input_size == (1022, 1022)
+
+
+@pytest.mark.parametrize("num_dataloader", [1, 2, 4])
+def test_get_adaptive_num_workers(mocker, num_dataloader):
+    num_gpu = 5
+    mock_torch = mocker.patch.object(target_file, "torch")
+    mock_torch.cuda.device_count.return_value = num_gpu
+
+    num_cpu = 20
+    mocker.patch.object(target_file, "cpu_count", return_value=num_cpu)
+
+    assert get_adaptive_num_workers(num_dataloader) == num_cpu // (num_gpu * num_dataloader)
+
+
+def test_get_adaptive_num_workers_no_gpu(mocker):
+    num_gpu = 0
+    mock_torch = mocker.patch.object(target_file, "torch")
+    mock_torch.cuda.device_count.return_value = num_gpu
+
+    num_cpu = 20
+    mocker.patch.object(target_file, "cpu_count", return_value=num_cpu)
+
+    assert get_adaptive_num_workers() is None
+
+
+@pytest.fixture()
+def fxt_dm_dataset() -> DmDataset:
+    dataset_items = [
+        DatasetItem(
+            id=f"item00{i}_0",
+            subset="train",
+            media=None,
+            annotations=[
+                Label(label=0),
+            ],
+        )
+        for i in range(1, 101)
+    ] + [
+        DatasetItem(
+            id=f"item00{i}_1",
+            subset="train",
+            media=None,
+            annotations=[
+                Label(label=1),
+            ],
+        )
+        for i in range(1, 9)
+    ]
+
+    return DmDataset.from_iterable(dataset_items, categories=["0", "1"])
+
+
+def test_get_idx_list_per_classes(fxt_dm_dataset):
+    # Call the function under test
+    result = get_idx_list_per_classes(fxt_dm_dataset)
+
+    # Assert the expected output
+    expected_result = defaultdict(list)
+    expected_result[0] = list(range(100))
+    expected_result[1] = list(range(100, 108))
+    assert result == expected_result
+
+    # Call the function under test with use_string_label
+    result = get_idx_list_per_classes(fxt_dm_dataset, use_string_label=True)
+
+    # Assert the expected output
+    expected_result = defaultdict(list)
+    expected_result["0"] = list(range(100))
+    expected_result["1"] = list(range(100, 108))
+    assert result == expected_result
+
+
+def test_import_object_from_module():
+    obj_path = "otx.data.utils.get_idx_list_per_classes"
+    obj = import_object_from_module(obj_path)
+    assert obj == get_idx_list_per_classes
